Move NavigationGUI Screen into separate include/, src/ files
